database part ca n be configured with default
