+++ src/Picums.Mvc/Configuration/Defaults/LocalisationByDatabase.cs
@@ -11,7 +11,7 @@ namespace Picums.Mvc.Configuration.Defaults
             host.Services.Add(x => x.AddTransient<GetAllTranslationsQuery>());
             host.Services.Add(x => x.AddTransient<ITranslationSetProvider, DataTranslationProvider>());
 
-            host.Apply<DataStorage>("Picums.Localisation");
+            host.Apply<CQRSDefaults>("Picums.Localisation");
         }
     }
 }
